requirements.txt
Flask
Flask-Migrate
flask-sqlalchemy
python-dotenv
psycopg2-binary
gunicorn
